// src/types/ordersType.ts
// Generated by https://quicktype.io

export interface OrderResponse {
  payment_uri: string;
  identifier: string;
  web_url: string;
  address: string;
  tag_memo: string;
  input_currency: string;
  expected_input_amount: number;
  rate: number;
  notes: string;
  reference: null;
  fiat: string;
  language: string;
}

// Generated by https://quicktype.io

export interface OrderInfo {
  identifier: string;
  reference: string;
  created_at: string;
  edited_at: string;
  status: string;
  fiat_amount: number;
  crypto_amount: number;
  unconfirmed_amount: number;
  confirmed_amount: number;
  currency_id: string;
  merchant_device_id: number;
  merchant_device: string;
  address: string;
  tag_memo: string;
  url_ko: string;
  url_ok: string;
  url_standby: string;
  expired_time: string;
  good_fee: boolean;
  notes: string;
  rbf: boolean;
  safe: boolean;
  fiat: string;
  language: string;
  percentage: number;
  received_amount: number;
  balance_based: string;
  internal_data: string;
  transactions: Transaction[];
}

export interface Transaction {
  confirmed: boolean;
  currency: string;
  amount: number;
  tx_hash: string;
  block: number;
  created_at: string;
}